+++ programs/pathfinder/src/instructions/withdraw.rs
@@ -143,9 +143,9 @@ pub fn process_withdrawal_and_transfer<'info>(
   accrue_interest(market, config)?;
 
   if *assets > 0 {
-    *shares = to_shares_up(*assets, market.total_deposits, market.total_shares)?;
+    *shares = to_shares_up(*assets, market.total_deposits as u64, market.total_shares)?;
   } else {
-    *assets = to_assets_down(*shares, market.total_deposits, market.total_shares)?;
+    *assets = to_assets_down(*shares, market.total_deposits as u64, market.total_shares)?;
   }
 
   // Update market total shares
@@ -170,7 +170,7 @@ pub fn process_withdrawal_and_transfer<'info>(
   // Update market total deposits
   market.total_deposits = market
     .total_deposits
-    .checked_sub(*assets)
+    .checked_sub(*assets as u128)
     .ok_or(error!(MarketError::MathUnderflow))?;
 
   // Transfer tokens
